+++ blueman/main/DBusProxies.py
@@ -51,7 +51,7 @@ class ManagerService(ProxyBase):
                          object_path="/org/blueman/Manager",
                          flags=Gio.DBusProxyFlags.DO_NOT_AUTO_START_AT_CONSTRUCTION)
 
-    def _activate(self) -> None:
+    def activate(self) -> None:
         try:
             param = GLib.Variant('(a{sv})', ({},))
             self.call_sync("Activate", param, Gio.DBusCallFlags.NONE, -1, None)
@@ -78,9 +78,3 @@ class ManagerService(ProxyBase):
 
         param = GLib.Variant('(sava{sv})', (name, [], {}))
         self.call('ActivateAction', param, Gio.DBusCallFlags.NONE, -1, None, call_finish)
-
-    def startstop(self) -> None:
-        if self.get_name_owner() is None:
-            self._activate()
-        else:
-            self._call_action("Quit")